backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
trafilatura==1.12.0
readability-lxml==0.8.1
pymupdf==1.24.0
pdfplumber==0.11.0
beautifulsoup4==4.12.3
transformers==5.3.0
torch==2.4.0
gliner==0.2.25
lingua-language-detector==2.0.2
sentencepiece==0.2.0
protobuf==5.27.0
ollama==0.3.0
nltk==3.9.1
python-multipart==0.0.9
accelerate==0.33.0
scipy==1.14.0
IndicTransToolkit
python-dotenv==1.0.1
